Finished for now
The circle isn't 100% right and it still doesn't work when the diameter is an even number but I have banged my head against the wall for a few hours and can't seem to solve it.

It wasn't really about the circle at the end of the day, this is just a project to help familiarize myself with JavaFX and to get back into coding after summer... or at least that's my coping mechanism. Maybe I'll come back and finish this another day?